+++ tests/faces/test_cylindricalface3d.py
@@ -108,7 +108,7 @@ class TestCylindricalFace3D(unittest.TestCase):
         self.assertTrue(face.surface2d.outer_contour.is_ordered())
         self.assertAlmostEqual(face.surface2d.area(), 0.003143137591511259, 3)
 
-    def neutral_fiber(self):
+    def test_neutral_fiber(self):
         face = self.cylindrical_face1
         neutral_fiber = face.neutral_fiber()
         self.assertEqual(neutral_fiber.length(), 0.4)
@@ -147,6 +147,56 @@ class TestCylindricalFace3D(unittest.TestCase):
             'faces/objects_cylindrical_tests/test_buggy_split_by_plane12_07_2023.json').primitives
         plane_intersections = face.plane_intersections(plane)
         self.assertAlmostEqual(plane_intersections[0].length(), 0.10485331158773475)
+
+    def test_conicalface_intersections(self):
+        expected_results = [[[3.710002373020358], [2.7546701820620956, 0.7935213268610651],
+                             [2.075126659615459, 0.49133092704442394, 1.0377142838361861, 0.5464208526645894],
+                             [2.5645345026338227, 2.5645345026338213], [0.544055470762666, 0.04555225339913864,
+                                                                        1.2782308617047107, 0.2561660954970303]],
+                            [[0.9041805829899517, 1.3797833276615086], [2.7546701820620956, 0.7935213268610651],
+                             [0.9945099057847868, 0.011885786063087272, 0.49133092704442394, 1.0377142838361861,
+                              0.5464208526645894],
+                             [0.28956380719439634, 0.9392209120999059, 2.5645345026338213],
+                             [0.2798809806923827, 0.045552253399138556, 0.7579657254386307]],
+                            [[0.8560429148647001, 0.32222894295393023], [0.6888878304141478, 0.6888878304141477,
+                                                                         0.1984154944167925, 0.19841549441679263],
+                             [0.49133092704442394, 1.0377142838361861, 0.5464208526645894],
+                             [2.5645345026338213],
+                             []]]
+        conical_surface = surfaces.ConicalSurface3D(volmdlr.OXYZ, math.pi / 6)
+        conical_face = faces.ConicalFace3D.from_surface_rectangular_cut(
+            conical_surface, 0, volmdlr.TWO_PI, 0, 2)
+
+        cylindrical_surface1 = surfaces.CylindricalSurface3D(volmdlr.Frame3D(volmdlr.Point3D(.3, .3, 0.8),
+                                                                             volmdlr.Y3D, volmdlr.Z3D, volmdlr.X3D),
+                                                             0.3)
+        cylindrical_surface2 = surfaces.CylindricalSurface3D(volmdlr.Frame3D(volmdlr.Point3D(0, 0, 0.5),
+                                                                             volmdlr.Y3D, volmdlr.Z3D, volmdlr.X3D),
+                                                             0.3)
+        cylindrical_surface3 = surfaces.CylindricalSurface3D(volmdlr.Frame3D(volmdlr.Point3D(0, 0, 1),
+                                                                             volmdlr.Y3D, volmdlr.Z3D, volmdlr.X3D),
+                                                             0.3)
+        point = volmdlr.Point3D(0, math.tan(conical_surface.semi_angle), 1)
+        normal = volmdlr.Vector3D(0.0, -0.5773502691896256, 1.0)
+        center = point + normal * math.tan(conical_surface.semi_angle) / 2
+        cylindrical_surface4 = surfaces.CylindricalSurface3D(volmdlr.Frame3D(center,
+                                                                             volmdlr.Y3D, volmdlr.Z3D, volmdlr.X3D),
+                                                             math.tan(conical_surface.semi_angle) / 2)
+        vector1 = volmdlr.Vector3D(1, 1, 1)
+        vector1 = vector1.unit_vector()
+        vector2 = vector1.deterministic_unit_normal_vector()
+        vector3 = vector1.cross(vector2)
+        frame = volmdlr.Frame3D(volmdlr.O3D, vector1, vector2, vector3)
+        cylindrical_surface5 = surfaces.CylindricalSurface3D(frame, math.tan(conical_surface.semi_angle) / 2)
+
+        for i, z in enumerate([-1, -0.5, 0]):
+            for j, cylindrical_surface in enumerate([cylindrical_surface1, cylindrical_surface2, cylindrical_surface3,
+                                                     cylindrical_surface4, cylindrical_surface5]):
+                cyl_face = faces.CylindricalFace3D.from_surface_rectangular_cut(
+                    cylindrical_surface, 0, volmdlr.TWO_PI, z, 2)
+                list_curves = cyl_face.face_intersections(conical_face)
+                for curve_solution, expected_result in zip(list_curves, expected_results[i][j]):
+                    self.assertAlmostEqual(curve_solution.length(), expected_result)
 
 
 if __name__ == '__main__':