+++ volmdlr/curves.py
@@ -3,6 +3,7 @@ Volmdlr curves.
 
 """
 import math
+import sys
 from typing import List, Union
 
 import matplotlib.pyplot as plt
@@ -26,6 +27,7 @@ class Curve(DessiaObject):
     """Abstract class for a curve object."""
 
     def __init__(self, name: str = ''):
+        self.periodic = False
         DessiaObject.__init__(self, name=name)
 
     def abscissa(self, point):
@@ -85,6 +87,9 @@ class Curve(DessiaObject):
 
 class ClosedCurve(Curve):
     """Abstract class for defining closed curves (Circle, Ellipse) properties."""
+    def __init__(self, name: str = ''):
+        Curve.__init__(self, name=name)
+        self.periodic = True
 
     def point_at_abscissa(self, abscissa):
         """
@@ -785,7 +790,7 @@ class Line3D(Line):
         :return: None if there is no intersection between Lines.
         A volmdlr.Point3D if there exists an intersection.
         """
-        return self.intersection(line)
+        return [self.intersection(line)]
 
     def plot(self, ax=None, edge_style: EdgeStyle = EdgeStyle()):
         """Plot method for Line 3D using Matplotlib."""
@@ -1528,14 +1533,15 @@ class Circle3D(CircleMixin, ClosedCurve):
         return start.rotation(self.frame.origin, self.frame.w,
                               curvilinear_abscissa / self.radius)
 
-    def line_intersections(self, line: Line3D):
+    def line_intersections(self, line: Line3D, abs_tol: float = 1e-6):
         """
         Calculates the intersections between the Circle3D and a line 3D.
 
         :param line: line 3D to verify intersections
+        :param abs_tol: Tolerance.
         :return: list of points intersecting Circle
         """
-        circle3d_line_intersections = volmdlr_intersections.circle_3d_line_intersections(self, line)
+        circle3d_line_intersections = volmdlr_intersections.circle_3d_line_intersections(self, line, abs_tol)
         return circle3d_line_intersections
 
     def linesegment_intersections(self, linesegment: 'volmdlr.edges.LineSegment3D', abs_tol: float = 1e-6):
@@ -2528,6 +2534,44 @@ class Hyperbola2D(HyperbolaMixin):
         points_positive_branch = [self.frame.local_to_global_coordinates(point) for point in points_positive_branch]
         return points_positive_branch
 
+    def point_belongs(self, point, abs_tol: float = 1e-6):
+        local_point = self.frame.global_to_local_coordinates(point)
+        if math.isclose(local_point.x**2/self.semi_major_axis**2 - local_point.y*+2/self.semi_minor_axis**1,
+                        1, abs_tol=abs_tol):
+            return True
+        return False
+
+    def get_dx_dy(self, point):
+        return (self.semi_major_axis**2 * point.y) / (self.semi_minor_axis**2 * math.sqrt(
+            self.semi_major_axis**2 * point.y**2 / self.semi_minor_axis**2 + self.semi_major_axis**2))
+
+    def tangent(self, point):
+        """
+        Calculates the tangent vector to an hyperbola at a given point.
+
+        :param point: The point at which the tangent vector is to be calculated.
+        :type point: volmdlr.Point2D.
+        :return: The tangent vector to the ellipse at the given point.
+        :rtype: volmdlr.Vector2D.
+
+        """
+        # Convert the point to local coordinates within the ellipse's frame
+        point_at_local_coord = self.frame.global_to_local_coordinates(point)
+
+        # Calculate the slope of the tangent line at the given abscissa
+        dy_dx = self.get_dx_dy(point)
+
+        # Construct the second point on the tangent line still on ellipse's frame.
+        tangent_second_point = point_at_local_coord + volmdlr.Point2D(dy_dx, 1)
+
+        # Convert the second point back to global coordinates
+        global_coord_second_point = self.frame.local_to_global_coordinates(tangent_second_point)
+
+        tangent_vector = global_coord_second_point - point
+        tangent_vector = tangent_vector.to_vector()
+
+        return tangent_vector
+
     def line_intersections(self, line: Line2D):
         """
         Calculates the intersections between a Hyperbola and an infinite Line in 2D.
@@ -2744,6 +2788,24 @@ class ParabolaMixin(Curve):
         """
         return 0.5 * (x ** 2) / (2 * self.focal_length)
 
+    # def trim(self, point1, point2):
+    #     """
+    #     Trims a Parabola between two points.
+    #
+    #     :param point1: point 1 used to trim circle.
+    #     :param point2: point2 used to trim circle.
+    #     """
+    #     _bspline_class = getattr(volmdlr.edges, 'BSplineCurve'+self.__class__.__name__[-2:])
+    #     local_split_start = self.frame.global_to_local_coordinates(point1)
+    #     local_split_end = self.frame.global_to_local_coordinates(point2)
+    #     max_x = max(local_split_start.x, local_split_end.x)
+    #     min_x = min(local_split_start.x, local_split_end.x)
+    #     parabola_points = self.get_points(min_x, max_x, 100)
+    #     if not parabola_points[0].is_close(point1):
+    #         parabola_points = parabola_points[::-1]
+    #     bspline = _bspline_class.from_points_interpolation(parabola_points, 2)
+    #     return bspline
+
     def trim(self, point1, point2):
         """
         Trims a Parabola between two points.
@@ -2751,16 +2813,15 @@ class ParabolaMixin(Curve):
         :param point1: point 1 used to trim circle.
         :param point2: point2 used to trim circle.
         """
-        _bspline_class = getattr(volmdlr.edges, 'BSplineCurve'+self.__class__.__name__[-2:])
-        local_split_start = self.frame.global_to_local_coordinates(point1)
-        local_split_end = self.frame.global_to_local_coordinates(point2)
-        max_x = max(local_split_start.x, local_split_end.x)
-        min_x = min(local_split_start.x, local_split_end.x)
-        hyperbola_points = self.get_points(min_x, max_x, 60)
-        if not hyperbola_points[0].is_close(point1):
-            hyperbola_points = hyperbola_points[::-1]
-        bspline = _bspline_class.from_points_interpolation(hyperbola_points, 2)
-        return bspline
+        _bspline_class = getattr(volmdlr.edges, 'BezierCurve' + self.__class__.__name__[-2:])
+        _line_class = getattr(sys.modules[__name__], 'Line'+ self.__class__.__name__[-2:])
+        tangent_vector1 = self.tangent(point1)
+        tangent_vector2 = self.tangent(point2)
+        lineseg1 = _line_class(point1, point1 + tangent_vector1)
+        lineseg2 = _line_class(point2, point2 + tangent_vector2)
+        line_inters = lineseg1.line_intersections(lineseg2)
+        bezier_parabola = _bspline_class(2, [point1, line_inters[0], point2])
+        return bezier_parabola
 
 
 class Parabola2D(ParabolaMixin):
@@ -2798,6 +2859,13 @@ class Parabola2D(ParabolaMixin):
             points.append(self.frame.local_to_global_coordinates(volmdlr.Point2D(x, y)))
         return points
 
+    def point_belongs(self, point, abs_tol: float = 1e-6):
+        local_point = self.frame.global_to_local_coordinates(point)
+        if math.isclose(local_point.y,
+                        self.vrtx_equation_a * local_point.x**2, abs_tol=abs_tol):
+            return True
+        return False
+
     def line_intersections(self, line: Line2D):
         """
         Gets intersections between a Parabola 2D and a Line 2D.
@@ -2821,6 +2889,49 @@ class Parabola2D(ParabolaMixin):
             x = m / (2 * self.vrtx_equation_a)
             return [volmdlr.Point2D(x, m * x + c)]
         return []
+
+    def tangent(self, point):
+        """
+        Calculates the tangent vector to a parabola at a given point.
+
+        :param point: The point at which the tangent vector is to be calculated.
+        :type point: volmdlr.Point2D.
+        :return: The tangent vector to the ellipse at the given point.
+        :rtype: volmdlr.Vector2D.
+
+        """
+        # Convert the point to local coordinates within the parabola's frame
+        point_at_local_coord = self.frame.global_to_local_coordinates(point)
+
+        # Calculate the slope of the tangent line at the point
+        dy_dx = 2 * self.vrtx_equation_a * point_at_local_coord.x
+
+        # Construct the second point on the tangent line still on parabola's frame.
+        tangent_second_point = point_at_local_coord + volmdlr.Point2D(1, dy_dx)
+
+        # Convert the second point back to global coordinates
+        global_coord_second_point = self.frame.local_to_global_coordinates(tangent_second_point)
+
+        tangent_vector = global_coord_second_point - point
+        tangent_vector = tangent_vector.to_vector()
+
+        return tangent_vector
+
+    def trim(self, point1, point2):
+        """
+        Trims a Parabola between two points.
+
+        :param point1: point 1 used to trim circle.
+        :param point2: point2 used to trim circle.
+        """
+        tangent_vector1 = self.tangent(point1)
+        tangent_vector2 = self.tangent(point2)
+        lineseg1 = Line2D(point1, point1 + tangent_vector1)
+        lineseg2 = Line2D(point2, point2 + tangent_vector2)
+        line_inters = lineseg1.line_intersections(lineseg2)
+        bezier_parabola = volmdlr.edges.BezierCurve2D(2, [point1, line_inters[0], point2])
+
+        return bezier_parabola
 
     def split(self, split_start, split_end):
         """Splits a Parabola between a start and end point."""
@@ -2917,6 +3028,13 @@ class Parabola3D(ParabolaMixin):
         v_vector = (v_point2d - origin).to_vector().unit_vector()
         frame = volmdlr.Frame2D(origin, u_vector, v_vector)
         return Parabola2D(frame, self.focal_length)
+
+    def tangent(self, point):
+        point_2d = point.to_2d(self.frame.origin, self.frame.u, self.frame.v)
+        tangent_2d = self.self_2d.tangent(point_2d)
+        point_tangent_2d = point_2d + tangent_2d
+        point_tangent_3d = point_tangent_2d.to_3d(self.frame.origin, self.frame.u, self.frame.v)
+        return (point_tangent_3d - point).to_vector()
 
     def point_belongs(self, point, tol: float = 1e-6):
         """